Moved files around to be consistent with file structure
