+++ src/Home/HomeScreen.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Text, View } from 'react-native';
 
-import TestFireBase from '../FireBase';
+import TestFireBase from '../Components/FireBase/Storage';
 
 export default function HomeScreen() {
